Ditch reconnect logic in favor of rebuilding dropped connections
We've narrowed down a memory leak issue to this driver, and this represents a
rather blunt solution to the problem.

My best guess is that it's possible for the Kafka connection to end up in a
disconnected state without any active reconnects (I can't reason how, but it
doesn't seem farfetched). In that case, sending data through the connection
appears to simply grow a buffer without limit.

This circumvents that possibility by simple re-creating the connection object
whenever a drop occurs.